lcitool: Set make parallelism dynamically
We're about to make some changes to lcitool that would make it
more annoying to get access to the number of install-time vCPUs
from 'lcitool update', so let's use getconf instead.

As a side effect, this also copes better with the situation where
the builder was installed with a certain number of vCPUs but the
amount has since been modified.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Erik Skultety <eskultet@redhat.com>