configs: Add the cloud-init.conf to be used with upstream machines
This config is usually applied (with the exception of NoCloud) by the
cloud metadata service and executed by the cloud-init service within
the guest, so as such it's not going to be used by any of the Ansible
tasks. However, the idea here is to publicly share the default config
that we're going to apply on our upstream machines.

Signed-off-by: Erik Skultety <eskultet@redhat.com>