osinfo-db: Add back intltool
Sadly itstool is not usable in the same way across all the supported
distributions of osinfo-db [1], so it is not possible to switch to it.
Hence, add intltool back as it is still used; leave itstool there, so
there is no need to add it again in the future in case we manage to get
itstool fixed.

[1] see:
https://gitlab.com/libosinfo/osinfo-db/-/merge_requests/133#note_342323692

Signed-off-by: Pino Toscano <ptoscano@redhat.com>