guests: don't force user to create a config.yaml file
Only the "install", "build" and "update" commands require a per-user
config.yaml with a password set. We should not require this for the
other commands to be run.

This splits up the validation into two parts. The first part which
always runs just removes unknown config keys. The second part is only
run for commands that deal with VMs, and checks for mandatory keys.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>