guests: Add CentOS Stream
The idea behind Stream is to get a preview of what the next RHEL-minor
will look like, thus getting newer versions of packages than regular
CentOS provides. It's supposed to be rolling, so currently it's based on
CentOS 8, but once RHEL-9 is out, it will point to the next RHEL-9
instead.  However, unlike debian-sid it doesn't have a dedicated URL
resource (something like centos-stream), so we'll have to change it
manually once it's flipped to point to next RHEL-9.

Signed-off-by: Erik Skultety <eskultet@redhat.com>