lcitool: Make some internal operations more generic
Both when removing unknown keys from the user-provided
configuration and when merging it into the default configuration,
we don't need to hardcode the list of sections to operate on,
since we can just look at the default configuration and extract
it from there. With this change, adding more sections will become
easier and less error-prone.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>