lcitool: Add MinGW capabilities to the Dockerfile generator
This is similar to the existing support for Linux cross-builds,
but in this case we can't programmatically figure out the name of
the foreign package; luckily, mappings are already defined in a
way that allows for architecture-specific entries, so we can simply
look for those when targeting MinGW.

The algorithm for Linux cross-builds is tweaked as well, but
that's just in order to keep it as similar as possible to the one
for MinGW cross-builds, and there are no behavioral changes.
Unfortunately, the two are still just different enough that we
can't merge them in a way that's maintainable.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>