mappings: Change cross-policy-deb to cross-policy-default
Whether you need to install the native package or the foreign one
mostly depend on whether it's an utility / build tool or a library,
which is something that doesn't really change across targets.

Let's use a more general key, which can always be overriden by a
more specific one if necessary.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>