lcitool: Remove check on host's projects
We have already defined a number of projects, like libvirt+dist,
that don't show up in this list. This is not just an oversight:
depending on the specifics of your environment, you might or might
not want those projects, and there is no right answer; the only
solution is for the user to pick exactly the projects they care
about, and specify them explicitly on the lcitool command line.

Basically, the idea of having a static list of projects that are
appropriate for a specific host has become obsolete, so let's stop
enforcing it.

The equivalent check has already been removed from the Dockerfile
generator in b71ad5e8257e.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>