guests: Drop +mingw* variants of all projects
Now that we are treating MinGW like all other cross-building
architectures, we no longer need to have special variants of the
various projects.

As a consequence of this change, it's no longer possible to use
'lcitool update' or 'lcitool build' on MinGW projects: while this
may appear to be a significant loss in terms of functionality,
we've been working towards moving all builds (including MinGW
builds and Linux cross-builds) to containers for a while now: the
plan going forward is to use VMs only for functional testing, and
in that context cross-builds are not relevant.

Jenkins jobs are unaffected, because they don't use lcitool but
have their own rules instead. As far as mingw* packages installed
on the existing CentOS CI builders are concerned, with this change
we lose the ability to add more, but they won't suddenly disappear
and will be kept up to date by the usual scheduled jobs until the
time comes, hopefully pretty soon, when we can finally shut down
that environment for good.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>