host_vars: Add a new path for pip
Currently when installing packages from pip, we use the generic Ansible
"command" module with "pip3" as the binary.
In the future, we'd like to switch to using the native Ansible "pip"
module. However, on FreeBSD there's no unversioned "pip3" binary, thus
the native module fails with:

"Unable to find any of pip3 to use.  pip needs to be installed."

Fix it by extending the "paths" variable for all the hosts to include
the correct pip3 path.

Signed-off-by: Erik Skultety <eskultet@redhat.com>