guests: Add macOS 10.15
Unsurprisingly, in a lot of cases this platform is very similar
to FreeBSD, but there are a number of ways they diverge.

A notable difference between macOS and all other platforms we've
supported so far is that Python and Perl modules are generally not
packaged, which means that we have to rely on PyPi and CPAN
respectively for those.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>