lcitool: Convert os.path to pathlib
os.path is a bit cumbersome in that it represents paths as strings,
which may not scale with the regular string methods when trying to
separate different segments of a path.
Pathlib on the other hand deals with this (and many other portability
challenges). This patch is necessary to help us split the main
lcitool script into individual modules.

Signed-off-by: Erik Skultety <eskultet@redhat.com>