lcitool: Initiate a new directory hierarchy for lcitool
The effort is to create a Python package out of lcitool which means to
start following some standard-ish directory hierarchy for the main
executable as well as the modules.
This patch introduces a new directory "lcitool" which will become a home
for everything lcitool related. The path to the "base" dir, aka guests/
had to be tweaked accordingly to keep everything working.

Signed-off-by: Erik Skultety <eskultet@redhat.com>