playbooks: build: libvirt: Don't run RPM job on CentOS
After libvirt switched to the meson build system, rpmbuild on CentOS is
no longer viable as meson >= 0.54.0 is required while only 0.49.2 (at
the time of this patch) is available.
Even if --nodeps is passed, the "%meson" RPM macro won't expand and the
build will fail.
This patch should be reverted once libvirt resolves this issue in their
specfile.

https://www.redhat.com/archives/libvir-list/2020-August/msg00262.html

Signed-off-by: Erik Skultety <eskultet@redhat.com>