removing get_base() in favor of pkg_resources.resource_filename
There are many ways to install a python package and we need to properly
distribute the 'data_files'. pkg_resources library has a
'resource_filename' utility method to help discovering those data files
when the package is installed on a system. Now that we are using this
for getting ansible/, cross/, etc/ and configs/ files, we could use it
on utils.py also instead of guessing the abs path of this tool.

Signed-off-by: Beraldo Leal <bleal@redhat.com>