Provide a script for running lcitool from git
Since the splitting of the lcitool code into python modules it is
impossible to actually run the tool without setting PYTHONPATH to point
to the modules. This is tedious when normal practice is to run the tool
straight from git to ensure you always have the latest logic / mappings.

This provides a wrapper script to make it possible to run lcitool
directly again without setting magic env vars.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>