config: Prefer the .yml extension over .yaml
All of the YAML files within the repo end with the '.yml' extension.
I neglected this fact back when I added support for a standalone config
file and used .yaml for it. So it gets a little confusing when working
with the files and copying the default config to HOME. In order to fix
this in a backwards compatible way as well as remain consistent within
the repo wrt YAML naming, accept both variants for the config - '.yaml'
and '.yml'.

Signed-off-by: Erik Skultety <eskultet@redhat.com>