host_vars: use distro provided container registry where available
Docker Hub has started to restrict the number of pulls that users can
make from its registry. In additional the content on Docker Hub is often
outdated compared to the distros' own builds. For example Fedora Rawhide
containers on Docker Hub can be as much as a month or more out of date
because every change to the image has to go through a pull request and
review process. The Fedora registry for rawhide meanwhile is updated as
often as once a day.

Fedora, CentOS and OpenSUSE all provide their own container registry
which does not have usage restrictions, and is guaranteed to hold the
most recent content.

The main downside of this change is that we'll be triggering rebuilds of
the libvirt containers more frequently for Fedora rawhide at least.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>