guests: introduce osinfo-db-tools+dist project
The libosinfo Mingw CI jobs previously attempted todo a native build of
osinfo-db-tools, followed by a mingw build of libosinfo.

The container, however, is not populated with packages needed to do the
native build of osinfo-db-tools, since the way lcitools handles mingw
was aligned with other cross builds.

The simplest solution is to just use the pre-built distro
osinfo-db-tools for mingw builds.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>