guests: introduce a "lcitool manifest" command
This command is intended to replace the various "refresh" shell scripts
in each project.

A "ci/manifest.yml" file in the project's git repo defines the desired
CI characteristics, including which hosts to generate containers for,
which cross builds to perform, which cirrus builds to perform. It then
generates the docker recipes, cirrus vars files and most of the gitlab
CI rules.

The developer just has to provide the project specific parts of the
build process themselves.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>