update: base: Fix the CentOS-Stream PowerTools repo naming regression
Just like it happened for CentOS-8, the same happened for Stream as
well. However, this time it's more complicated, not only the repo name
changed, but the original CentOS-Stream-* repos were dropped and
instead the original CentOS-Linux-* repos updated and now they point to
the Stream content.
All of ^this was incorporated to a new package 'centos-stream-release'
which now replaces the original 'centos-release-stream' one, except
unlike its predecessor it cannot be installed from the Extras repo, so
the proper way how Stream upgrade needs to be handled for this
regression is:
install centos-release-stream-> install centos-stream-release-> enable
the powertools repo-> update the rest of packages

If you enable the PowerTools repo first then either upgrade or install
the centos-stream-release package, you're left with an .rpmsave config
in /etc/yum.repos.d which not even rpmconf knows what to do with.
Simply put, the result was the e.g. 'ninja-build' could not be found
and installed.

Signed-off-by: Erik Skultety <eskultet@redhat.com>