formatters: Install the Stream repos for CentOS first, then update
Normally, the logical order of operations is update->package install.
For CentOS-Stream however, we first need to install the Stream repos in
order to enable the Stream contents and only then can we update the
system packages, otherwise we may end up with some packages in Stream
version and some in the base CentOS-8 versions (unless a specific
Stream package install enforces an explicit upgrade).

Signed-off-by: Erik Skultety <eskultet@redhat.com>