application: Use a global temporary directory
We don't need to create multiple temporary ones for a single run of the
application. Future patches will convert our direct Ansible usage to
using the ansible-runner package for which we'll also create some
files to contain the Ansible execution environment.
For that reason it would be beneficial if we had a single global
temporary directory which would live throughout a single lcitool run
and which could host data for various use cases.

This patch introduces a util helper that initializes the temporary
directory only once and behaves basically like a singleton.

Signed-off-by: Erik Skultety <eskultet@redhat.com>