lcitool: Introduce AnsibleWrapper class
This class wraps our usage of Ansible. More specifically, it wraps the
Runner and RunnerConfig classes which come from the ansible-runner
package [1].
The wrapper takes care of the following:
    - provides API entrypoints for common Ansible executions, thus
      abstracting the specifics of both RunnerConfig and Runner objects
    - prepares an environment for the actual Ansible runner to run in
      (each Ansible runner needs a private data dir, e.g.
       inventories, playbooks etc.)

The benefit of using ansible-runner interface is that it's meant to be
stable and wraps shell executions of Ansible properly, e.g. by
providing return codes, stats, stdout output, etc. It's also event
based should we fancy using a complex event loop mechanism.

[1] https://ansible-runner.readthedocs.io/en/stable/index.html

Signed-off-by: Erik Skultety <eskultet@redhat.com>