util: Introduce a helper for key lookups in nested dicts
This helper will become relevant once we let Ansible handling the
inventories for us. In other words, in order to extract a full list of
hosts out of (potentially) multiple inventory sources, we need Ansible
to process this for us and return the resulting inventory in a format
we can consume back - YAML. The thing is that the nesting of the YAML
dictionary is fairly deep and so if we need to extract all the hosts
from all the groups, we need to traverse the whole dictionary to find
all occurrences of the 'hosts' and fetch the data.
Since YAML is no XML Dom, we need our own recursive helpers returning
a generator to the data in question.

Signed-off-by: Erik Skultety <eskultet@redhat.com>