lcitool: Replace '.format()' occurrences with f-strings
f-strings were introduced in Python 3.6 as a new (and faster) way of
formatting parametrized strings that are supposed to lead to a more
concise code.

This patch replaces all occurrences of <str>.format() except for the
code bits working with the 'varmap' dictionary under the formatters
module - consistency is one thing, but in this case building a giant
list of strings and then calling a single '.format()' in order to
populate the values is much better than fiddling with f-strings
gradually.

Converting to f-string usage will help us down the road when we'll
replace the way we use exceptions (most notably re-raising) with proper
error reporting through the logging library.

Signed-off-by: Erik Skultety <eskultet@redhat.com>