logging: Set up a global logger instance
Exceptions are a powerful error handling mechanism, but the way we use
them is less than ideal. They're targeted at code writers not users
of the project since it's essentially a debugging tool for reporting
details about errors coming either from the Python runtime or a
library.

Python standard library comes with a logging module which makes proper
logging a piece of cake. In most cases it's enough to setup the "root"
logger with the desired logging level (default is warning), setting
the error message format and providing the output stream where
the messages should be sent. Once set up, the rest of the code needs to
obtain a reference to the global logger object and use it to start
logging messages of all levels.

Unfortunately for us, we need a bit more than just the root logger with
its default settings. That's because the stock formatter can only
handle a single format at a time. We'd like to keep the error level
message returned to the user compact, but at the same time be more
verbose about the origin of a message with debug level messages.
For that, we can either use 2 logger instances or we use a custom
Formatter subclass which will track the logging level of each message
and select a proper formatter instance accordingly. Since the latter
is more straightforward this patch implements that one.

Signed-off-by: Erik Skultety <eskultet@redhat.com>