application: Introduce an expand pattern helper
The problem with the expand_pattern utility function is that it raises
exceptions depending on what is wrong with the expansion pattern. That
means that in order to properly handle them we'd have to wrap every
expand_pattern call with a try-except block.
Instead, this patch introduces an Application level helper for pattern
expansion that does the wrapping and exiting if an exception occurred.

The other approach would be to just wrap the expand_pattern public
methods in the respective Inventory and Project modules, but if we ever
consider turning this project into a library, it is much better to
handle the exceptions in a single place and let all the other
exceptions be propagated from the rest of the code to the application
code or, in the future, to the library consumer.

Signed-off-by: Erik Skultety <eskultet@redhat.com>