util: Extract Dockerfile header generation to the util module
The plan is to move the code out of the Formatter module completely,
this patch is just preparatory work easy to review.
Since the original code used the old-style Python formatting mechanism
'%', this patch replaces it with the more modern ".format()" method
as well as makes use of triple quotes instead of an array of strings.

Signed-off-by: Erik Skultety <eskultet@redhat.com>