application: Generate the automatic header in the application code
The automatic header we generate for Dockerfiles (and now for
variables) is not really related to the formatting code, it's something
extra that the application adds, but strictly speaking the formatter
module should only ever return either a Dockerfile or variables as
output, no more, no less.
Also, the code should be symmetric for both
Dockerfiles and variables, so this patch enables the header also for
variables files.

Signed-off-by: Erik Skultety <eskultet@redhat.com>