mappings: Skip Clang when cross-building
Clang supports cross-compilation, but not in the way we expect
to use it: specifically, our meson cross-file explicitly points
to gcc, and MinGW is also built on top of gcc.

So, while it might be possible to use Clang for at least Linux
cross-builds, we're not really anywhere close to enabling that
use case at the moment, and so installing it inside containers
that are meant from cross-building only adds to their size.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>