lcitool: Change how we deal with cross-compilers
At the moment, cross-compilers are treated as very special cases:
instead of looking up the package name in the mappings file, it is
generated dynamically based on hard-coded information, and then the
resulting package is simply added to the list of cross_pkgs.

As it turns out, this special-casing is already not needed for
MinGW cross-compilers: the mapping file contains the appropriate
mappings for gcc, and so the changes made by the script are
effectively a no-op.

Modify the handling of Linux cross-compilers so that it works
similarly: the package names are now stored in the mappings file
along with all other names, and the only compiler-specific logic
is that we know not to append the foreign architecture name when
installing them.

This more consistent handling allows us to improve how we
create ccache wrappers: in addition to gcc, ccache wrappers are
now also created for clang and g++ if necessary.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>