mappings: Bump Meson to 0.56.0
Meson 0.55.3, which we're currently using, is affected by a bug
that causes it to fail early when a native compiler is not
available on the system - even when performing a cross-build.

Now that we're no longer installing the native gcc in containers
that are targeting cross-building, this results in

  meson.build:1:0: ERROR: Unknown compiler(s): ['cc', 'gcc', 'clang', 'pgcc', 'icc']
  The follow exceptions were encountered:
  Running "cc --version" gave "[Errno 2] No such file or directory: 'cc': 'cc'"
  Running "gcc --version" gave "[Errno 2] No such file or directory: 'gcc': 'gcc'"
  Running "clang --version" gave "[Errno 2] No such file or directory: 'clang': 'clang'"
  Running "pgcc --version" gave "[Errno 2] No such file or directory: 'pgcc': 'pgcc'"
  Running "icc --version" gave "[Errno 2] No such file or directory: 'icc': 'icc'"

Note that this doesn't change the list of targets where we
install Meson from PyPi, even though in theory we would have to
do that. The reason for this is that targets where we don't
perform cross-builds, like for example Fedora 33, can still
build everything just fine with their current version of Meson.
Basically we only need this for Debian 10.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>