mappings: Use perl-base instead of perl where possible
Debian-based distros, as well as Fedora and openSUSE, offer the
possibility to install a minimal version of Perl that is still
featureful enough to run things like autotools, so let's use
that where it's available to save some disk space.

The mapping is a bit more verbose than it would be strictly
necessary because it's written under the assumption that
changes made to Fedora will eventually show up in CentOS: as
time goes by, we should be able to drop most of the release
specific overrides, starting with Fedora 32 and CentOS 7 in
just a few months.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>