ansible: Move host_vars to group_vars
Each of the current host_vars directory doesn't describe just a single
host, but actually a family of hosts. If we want to support arbitrary
user hosts, we'll still need to apply the facts to their hosts as well.
Once we move hosts_vars to group_vars, the only necessary action from
the user's perspective eventually will be to place each of their hosts
under the corresponding target OS group. This will ensure that our
internal facts will be loaded for that host automatically.

Note that even though the code base is using group_vars instead of
host_vars everywhere, the Ansible runner is still reading the facts out
of the 'host_vars' directory, because we're still using our in-repo
static inventory which has no groups and so 'lcitool update' would be
broken. This fact will be corrected in a later patch when we completely
switch to user provided inventories.

Signed-off-by: Erik Skultety <eskultet@redhat.com>