inventory: Don't parse the inventory file and load host vars directly
Currently we're parsing our simple static inventory only to populate
the keys for our target OS facts cache. We'll drop the static inventory
eventually, so we should approach populating the facts cache
differently - we should use the name of each directory under
ansible/host_vars as the key to the cache and then take the parsed
main.yml file from that directory and use it as data for that specific
key.

Signed-off-by: Erik Skultety <eskultet@redhat.com>