ansible_wrapper: Replace ansible.cfg with environment variables
Ansible runner doesn't natively support ansible.cfg. In fact, all the
settings to the underlying Ansible process are supposed to be passed as
environment variables. Since the idea is to generate as much Ansible
stuff as possible and store the least amount of Ansible bits necessary
in the repo, we technically don't need ansible.cfg anymore.

In order to be able to get rid of ansible.cfg, this patch utilizes the
execution environment helper methods from the previous patch and makes
sure all the input data Ansible needs to work with is in place before
the actual execution of Ansible.

Signed-off-by: Erik Skultety <eskultet@redhat.com>