ansible_wrapper: Add methods to prepare environment for the runner
We already enabled usage of the Ansible runner in lcitool, but we still
relied on ansible.cfg to configure Ansible.
Instead, the way Ansible runner is supposed to work is that all the
input data should be copied to Ansible runner's runtime directory.
This patch follows that philosophy and introduces a helper method that
prepares the execution environment for the Ansible process by creating
the necessary directory hierarchy. Depending on whether we need to
touch the input data itself before executing Ansible the logic either
copies or symlinks the source data (playbooks, inventories, vars) to
the Ansible runner's runtime directory,

Signed-off-by: Erik Skultety <eskultet@redhat.com>