lcitool: Install CentOS Stream 8 from proper media
Back when CentOS Stream was introduced, it was handled as a
sort of add-on to CentOS 8; these days, CentOS Stream 8 is a
standalone OS, with proper installers and container images.
Let's use those.

As a consequence of this change, we no longer need to manually
enable additional release packages after installation.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>