lcitool: Change version name for CentOS Stream 8
Technically speaking, CentOS Stream is a separate OS from
CentOS, so this should really be

  os:
    name: 'CentOSStream'
    version: '8'

However, since most of the mappings for CentOS also apply to
CentOS Stream, pretending that it's just a separate version
rather than a separate OS means we don't have to introduce a
lot of duplicated mappings.

Both CentOS 7 and CentOS 8 will be gone by the end of the
year anyway: once that happens, we can consider adjusting
the OS information to match reality with no downsides.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>