commandline: Introduce --name to the 'install' command
To avoid collisions of target OS group names with our default automatic
host naming during installations, mandate a name to be provided for any
machine being installed. This is achieved with a required '--name'
option to the 'install' command.
This completes the possibility to have multiple instances of the same
target OS co-existing in libvirt.

Signed-off-by: Erik Skultety <eskultet@redhat.com>