ansible: host_vars: Add 'target' variable
All the build jobs depend on whether a specific node's hostname is
among the supported list of machines capable AND/OR designated for a
given build job type.
Now, we have the name of the OS, we have the OS version and we also know
the node's hostname. However, once we allow user-provided hosts,
we won't be able to rely on hostnames anymore and we'll need rely on
something that is common for all the hosts - target OS group.
Because the target will be common for all hosts under the same target
OS group, it's better if we pass the variable this way rather than
adding it dynamically through code.

Note that simply using Ansible's builtin mechanism to do this

{{ hostvars[inventory_hostname].group_names[0] }}

is both less readable and also doesn't guarantee that group_names[0]
will actually point to one of our target OS groups with user-provided
hosts.

Signed-off-by: Erik Skultety <eskultet@redhat.com>