commandline: Adopt 'targetopt' CLI argument by the 'install' command
Just like dockerfile and variables generators should work with targets,
the same goes for 'install'. However, the thing with machine installs
is that we currently allow several host installations to happen
sequentially with only a single invocation of lcitool.
This posed a limitation in that only a single instance of target OS
could exist at a time and its name was hardcoded by us.

Since we'll eventually delegate the responsibility of providing a
machine inventory to users, we should not honor this limitation anymore
and allow multiple instances of the same target OS to exist. For that
however, we must only allow a single target to be installed at a time.
Next patch will add a '--name' option to the 'install' command so that
users can also name the instances the way they want so that multiple
instances of the same target can co-exist in the libvirt ecosystem.

Signed-off-by: Erik Skultety <eskultet@redhat.com>