ansible: cfg: Place the log outside of the repo
Previous patch made sure that lcitool creates an app directory for
non-essential data. Since lcitool is not supposed to be a system app,
we don't have write permissions to /var/log, thus we put the Ansible
execution logs in the cache dir mentioned above, it's better to place
dynamic data outside of the main repo anyway.

Note that this patch hardcodes the path to "~/.cache" even though
previous patches enabled usage of $XDG_CACHE_HOME variable as well
which might point elsewhere. However, this is fine for now as later
patches which will drop ansible.cfg entirely.

Signed-off-by: Erik Skultety <eskultet@redhat.com>