containers: handle checking DCO on merge requests
When the check-dco script is triggered from a merge request pipeline,
rather than a branch push pipeline, there might be a merge commit
present. This will not have a SoB line present and thus will fail
the DCO check, despite all the real commits having valid SoB.

We cannot just check all commits from the ancestor. We need to use
the explicit ancestor and head from the merge request, which will
exclude the merge commit.

This also has the benefit that we no longer need to add an extra
remote for the upstream repository in the merge request scenario,
since we are no longer trying to guess the common ancestor.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>