util: Don't sort intermediary results
Sorting should happen right before returning the results to the end
user or dumping the results to a file, because in this case it also
resulted in an implicit type conversion. Sorting takes time and if
the direct consumer of the expanded results wishes to further process
the results they'll have re-sort again.

Note that util.expand_pattern will now return a set, which is
fine as the callers can cast to a list when done processing the
data.

Signed-off-by: Erik Skultety <eskultet@redhat.com>