application: Drop the _expand_pattern helper
This helper was created so that all pattern expansion were both
called in a module agnostic way as well as to avoid having to wrap
all invocations with a try-except block. Thanks to module-level
exceptions that are being caught in the Application.run() method
this helper is no longer needed.

Signed-off-by: Erik Skultety <eskultet@redhat.com>