inventory: Lazy load the inventory
Unlike with lcitool configuration, most commands actually do need the
inventory. However, at some point in the future we're going to
distinguish between abstract OS targets and actual hosts that can be
managed with lcitool. What this means is that we'll have to call into
Ansible to parse the inventory sources for us - this is operation is
expensive and most commands actually only need the list of targets,
not the list of hosts in the inventory, hence the inventory will need
to be lazy evaluated.
This patch prepares the code base for that switch.

Signed-off-by: Erik Skultety <eskultet@redhat.com>