projects: Extract project package list loading to a dedicated method
In its current shape this change is mostly to remain consistent with the
inventory and config modules which have separate methods for loading data.

As for the confusing name '_load_projects' - future patches will extract
more code from this method and loading of the packages themselves will
move to a new dedicated 'Project' class. As a result, the Projects class
will only hold an iterable structure filled with Project instances,
hence the name '_load_projects'.

Signed-off-by: Erik Skultety <eskultet@redhat.com>