config: Change the way we override the default config with user values
We store the resulting config in the 'values' instance attribute which
we directly change using the private _update() method. Future patches
will switch to a lazy evaluating code which will fill in the 'values'
attribute on demand. Therefore, after this patch the _update() method
stops working with the public attribute directly and instead takes 2
distinct arguments:
    - the default config and
    - user config
and returns a new dictionary which will then be assigned to the public
'values' attribute.
Because the signature of the original _update() method changes with
this patch and also because it now behaves slightly differently,
convert it to a static method and rename to something more
verbose --> _override_config().

Signed-off-by: Erik Skultety <eskultet@redhat.com>