projects: Extract the package mappings loading code to the Project class
Each project has a list of generic package names which will in turn be
translated to a list of actual package names - this structure better
conforms to the OOP principle and will help us later when extracting
the generic-to-actual package name resolver code to a dedicated class.

This patch extracts loading of the per-project generic package names
list to the Project class and in turn the Projects class now
initializes only a dictionary of Project instances, each such instance
pointing to its own list of per-project abstract package names.

Follow up patches will make sure the packages list is lazy loaded, so
when everything is finally in place, packages will only be loaded
through the Project class, only when needed and only projects
which need to be evaluated will be loaded, nothing else.

Signed-off-by: Erik Skultety <eskultet@redhat.com>