projects: Lazy load per-project generic packages list
Up until now we always loaded the generic packages lists for all
projects when the application was initialized. Since not all commands
need to work with either projects or packages, everything is now lazy
loaded on demand, so when the application is initialized, a dictionary
of Project() placeholders is stored in the Projects singleton instance
and only when a specific project and its packages need to be touched are
the actual data loaded from disk.

Signed-off-by: Erik Skultety <eskultet@redhat.com>