projects: Start resolving the package names by using the package factory
Now that we have a package factory class, we can start producing
instances of the Package class which already have their generic mapping
name resolved to the real-world package name.

Since there's no need for the Projects class to differentiate between
all the different mapping types, we only load the YAML mappings file
and pass the contents to the package factory code which will take care
of the rest.
Note that the logic introduced here will still not be enabled in the
formatters module after this patch and hence we need to keep all the
X_mappings class properties alive.
However, because the PackageFactory class needs to be initialized with
the complete mappings dictionary, we need to turn off the mapping
lazy loading for the next few patches.

Signed-off-by: Erik Skultety <eskultet@redhat.com>