application: Rename hosts and projects parameters of _execute_playbook
Next patch will need to introduce a pointer to the Projects() singleton
instance which should naturally be called "projects", but that would
collide with the one of _execute_playbook method's parameters. Besides,
what we get in the "projects" argument is actually a potential glob
pattern that we need to expand anyway.

Signed-off-by: Erik Skultety <eskultet@redhat.com>