.gitignore: Include the Ansible's playbook build directory back
It is a common practice to ignore all 'build/' directories as those are
the standard way for many toolchains to create as a runtime directory.
However, the 'build' directory which hosts our Ansible build playbooks
is a valid directory which we want to track and potentially add stuff
to, hence we need to include it back.

Signed-off-by: Erik Skultety <eskultet@redhat.com>