inventory: Let Ansible handle all the inventory sources
Right now when users issue the 'hosts' command, we parse our simple
INI-style static inventory to extract the list of hosts, match a subset
with a pattern and return to the user.
The problem arises when users will be able to supply their own
inventories - these may very well be dynamic inventories which are
scripts yielding results in a format that Ansible mandates. We surely
don't want to run user scripts to get the list of hosts, especially not
from multiple inventory sources. Therefore, let Ansible handle all of
the inventory processing, producing a YAML-formatted inventory as a
result which we can then take and process it further to extract the
list of hosts.

There's a big caveat though. If the user provides an inventory where
a group name matches a host name, Ansible emits a warning. Normally,
this would not be an issue in shell because the error output could be
redirected, but this is not possible with Ansible runner - there's only
a single output, stdout. Naturally, these extra warnings in the
returned "data" would prevent a YAML parser to accept it. So, the
solution is to register an event callback to the Ansible runner object,
process every event coming from the underlying Ansible process, filter
out the warnings and only keep the data.
Why is it okay to filter out the warnings?
    1) We're talking about ansible-inventory - basically all the
       warnings relate to potential ansible-playbook invocations that
       would follow.
    2) Because ansible-inventory returns "0 (success)" even with the
       warnings present so we have no idea of telling that there's
       something wrong with the inventory beforehand to be able to
       catch this.

Signed-off-by: Erik Skultety <eskultet@redhat.com>